app/requirements.txt
fastapi>=0.80.0
pandas>=1.3.0
xgboost>=2.0.0
uvicorn>=0.15.0
scikit-learn==1.3.2
python-multipart>=0.0.6
google-cloud-storage>=1.26.0,<2.0.0dev
joblib~=1.0
gcsfs>=2024.2.0